Split manager tests into sibling file to respect size gate (#200)
manager.rs crossed the 1000-line source-file hard limit (1024) after the
prefix-passthrough remediation. Extract the #[cfg(test)] module into
manager_tests.rs, matching the established commands.rs/commands_tests.rs
and tmux_driver.rs/tmux_driver_tests.rs pattern. No behavior change.
